Remove EncodeCall trait
